Changed system manager so charge doesnt decreease. Also fixed a bug where the enemies did no damage

+++ damageState.h
@@ -34,7 +34,7 @@ private:
     double engineHealth;
     double hullHealth;
 
-    int damageSustained;
+    double damageSustained;
 
     CollaborationInfo *pilotInfo;
     CollaborationInfo *engineerInfo;
